refactor and reuse setAttributesDefaultValues in enrichment
