+++ services/apps/data_sink_worker/src/service/memberAttribute.service.ts
@@ -1,3 +1,4 @@
+import { setAttributesDefaultValues } from '@crowd/common'
 import { DbStore } from '@crowd/data-access-layer/src/database'
 import {
   getMemberAttributeSettings,
@@ -22,47 +23,7 @@ export default class MemberAttributeService extends LoggerBase {
     attributes: Record<string, unknown>,
   ): Promise<Record<string, unknown>> {
     const priorities = await getPlatformPriorityArray(dbStoreQx(this.store), tenantId)
-    if (!priorities) {
-      throw new Error(`No priorities set for tenant '${tenantId}'!`)
-    }
-
-    for (const attributeName of Object.keys(attributes)) {
-      if (typeof attributes[attributeName] === 'string') {
-        // we try to fix it
-        try {
-          attributes[attributeName] = JSON.parse(attributes[attributeName] as string)
-        } catch (err) {
-          this.log.error(err, { attributeName }, 'Could not parse a string attribute value!')
-          throw err
-        }
-      }
-      const highestPriorityPlatform =
-        MemberAttributeService.getHighestPriorityPlatformForAttributes(
-          Object.keys(attributes[attributeName]),
-          priorities,
-        )
-
-      if (highestPriorityPlatform !== undefined) {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        ;(attributes[attributeName] as any).default =
-          attributes[attributeName][highestPriorityPlatform]
-      } else {
-        delete attributes[attributeName]
-      }
-    }
-
-    return attributes
-  }
-
-  public static getHighestPriorityPlatformForAttributes(
-    platforms: string[],
-    priorityArray: string[],
-  ): string | undefined {
-    if (platforms.length <= 0) {
-      return undefined
-    }
-    const filteredPlatforms = priorityArray.filter((i) => platforms.includes(i))
-    return filteredPlatforms.length > 0 ? filteredPlatforms[0] : platforms[0]
+    return setAttributesDefaultValues(tenantId, attributes, priorities)
   }
 
   public async validateAttributes(
