+++ services/apps/premium/members_enrichment_worker/src/activities.ts
@@ -6,6 +6,7 @@ import {
   getEnrichmentData,
   getEnrichmentInput,
   getObsoleteSourcesOfMember,
+  getTenantPriorityArray,
   insertMemberEnrichmentCache,
   isCacheObsolete,
   isEnrichableBySource,
@@ -71,4 +72,5 @@ export {
   squashMultipleValueAttributesWithLLM,
   squashWorkExperiencesWithLLM,
   updateMemberUsingSquashedPayload,
+  getTenantPriorityArray,
 }
