+++ services/libs/common/src/index.ts
@@ -25,5 +25,6 @@ export * from './emailDomainValidator'
 export * from './displayName'
 
 export * from './i18n'
+export * from './member'
 
 export { Error400, Error401, Error403, Error404, Error405, Error409, Error500, Error542 }
